nostr/requirements.txt
websockets>=12.0
near-api-py>=0.1.0
nostr-sdk>=0.1.0
